+++ CHANGELOG.md
@@ -8,13 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## 0.21.2- 2023-04-05
+## 0.21.3 - 2023-04-06
+
+### Added
+* ENTER as btn action for resume in debug
+* Support multi middle views
+* Pipeline data view sample
+
+
+## 0.21.2 - 2023-04-05
 
 ### Changed
 * Try to fix mutation problem and update LIST_STRING
 
 
-## 0.21.1- 2023-04-04
+## 0.21.1 - 2023-04-04
 
 ### Changed
 * LIST_STRING rework field format